+++ test/kiln_cms/org_settings_test.exs
@@ -145,6 +145,7 @@ defmodule KilnCMS.OrgSettingsTest do
       KilnCMS.CMS.SiteLocaleSettings,
       KilnCMS.CMS.SiteMailRelay,
       KilnCMS.CMS.SiteMeilisearch,
+      KilnCMS.CMS.SiteSsoProvider,
       KilnCMS.CMS.SiteStorage,
       KilnCMS.CMS.SiteVapidKey,
       KilnCMS.Federation.SiteFederation